fix warning
git-svn-id: svn://sporkhack.com/nh@36 0b9a868d-f130-0410-bc2a-adc7ae8015a6

+++ include/hack.h
@@ -101,7 +101,7 @@ NEARDATA extern coord bhitpos;	/* place where throw or zap hits or stops */
 #define INVIS_BEAM	4
 
 #define MATCH_WARN_OF_MON(mon)	 (Warn_of_mon && flags.warntype && \
-		   		 (flags.warntype == (mon)->data->mlet))
+		   		 (flags.warntype == (unsigned long)(mon)->data->mlet))
 
 #include "trap.h"
 #include "flag.h"
